Revisão códigos hierarquia grafo; correção nomes; compila; roda o grafo; roda calculo tortuosidade; falha calculo permeabilidade; revisar M6;

+++ src/EstruturaDados/ObjetoRede/CObjetoRede_Tipo.cpp
@@ -163,7 +163,7 @@ Lista_dos_rotulos_das_conexões
    out.setf( ios::right );
 
    // Tipo de contorno
-   out << setw( 5 ) << static_cast<uint8_t>( Tipo() ) << '\n';
+   out << setw( 5 ) << static_cast<uint16_t>( Tipo() ) << '\n';
 
    // Rótulo de this
    out << ' ' << setw( 5 ) << rotulo;
